Introduce a splash screen for an android version
On android there is a small time delay between the application
start and the first frame of the UI being drawn.
Since QT on android starts as a vanilla java application, it
displays an empty activity while starting the QT app.
Make the activity display a splash screen of the application.