Refactor CMakeLists.txt to properly build on Windows
Remove search directory from find_program() call:
it finds the program without the argument and the
argument was not working in the first place.
Add version of the original TodoList target for
WIN32 platform.
Define path, used in *deployqt program call, separately
for MacOS and Windows:
- macdeployqt needs to be pointed at the *.app directory
- windeployqt needs to be pointed at the directory,
containing the actual executable binary
In order to run debug target on windows, following
folders should be in PATH env variable:
- <Qt installation folder>/bin
- <Qt installation folder>/plugins/platforms