Implement initial version of Android compilation
Add a submodule, that defines a macro, used to compile
an android target.
Move all android sources to the android/sources
directory.
In addition to all the normal environment variables,
to build an .apk you need to specify additional
following CMake command line arguments:
-DCMAKE_TOOLCHAIN_FILE
-DCMAKE_FIND_ROOT_PATH_MODE_PACKAGE=ON
-DANDROID_PLATFORM=26
The compiled apk is only for a single platform at
the moment.